+++ packages/oauth/oauth-provider-frontend/src/data/useFriendlyClientId.ts
@@ -1,31 +1,24 @@
 import { useLingui } from '@lingui/react/macro'
+import {
+  isConventionalOAuthClientId,
+  isOAuthClientIdLoopback,
+} from '#/util/oauth-client'
 
 export function useFriendlyClientId({
   clientId,
-  clientTrusted = false,
 }: {
   clientId: string
-  clientTrusted?: boolean
 }): string {
   const { t } = useLingui()
 
-  if (clientId.startsWith('http://')) {
+  if (isOAuthClientIdLoopback(clientId)) {
     return t`loopback`
   }
 
-  if (clientId.startsWith('https://')) {
-    try {
-      const url = new URL(clientId)
-      if (clientTrusted) {
-        return url.hostname
-      }
-      if (url.pathname === '/oauth-client-metadata.json' && !url.port) {
-        return url.hostname
-      }
-    } catch {
-      // ignore
-    }
+  if (isConventionalOAuthClientId(clientId)) {
+    return new URL(clientId).hostname
   }
 
+  // Should never happen
   return clientId
 }