+++ packages/oauth/oauth-provider-frontend/src/routes/__root.tsx
@@ -1,4 +1,4 @@
-import { Outlet, createRootRoute } from '@tanstack/react-router'
+import { HeadContent, Outlet, createRootRoute } from '@tanstack/react-router'
 import { RouterErrorComponent } from '#/components/ErrorScreen'
 import { Footer } from '#/components/Footer'
 import * as Layout from '#/components/Layout'
@@ -11,6 +11,8 @@ export const Route = createRootRoute({
 function Root() {
   return (
     <>
+      <HeadContent />
+
       <Layout.Outer>
         <Outlet />
       </Layout.Outer>