+++ packages/ozone/src/lexicon/types/app/bsky/notification/listNotifications.ts
@@ -67,7 +67,7 @@ export interface Notification {
   uri: string
   cid: string
   author: AppBskyActorDefs.ProfileView
-  /** Expected values are 'like', 'repost', 'follow', 'mention', 'reply', 'quote', and 'starterpack-joined'. */
+  /** Expected values are 'like', 'repost', 'follow', 'mention', 'reply', 'quote', 'starterpack-joined', 'verified', and 'unverified'. */
   reason:
     | 'like'
     | 'repost'
@@ -76,6 +76,8 @@ export interface Notification {
     | 'reply'
     | 'quote'
     | 'starterpack-joined'
+    | 'verified'
+    | 'unverified'
     | (string & {})
   reasonSubject?: string
   record: { [_ in string]: unknown }