+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,66 @@
 # @atproto/dev-env
 
+## 0.3.125
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/pds@0.4.132
+  - @atproto/bsky@0.0.144
+  - @atproto/ozone@0.1.106
+
+## 0.3.124
+
+### Patch Changes
+
+- [#3773](https://github.com/bluesky-social/atproto/pull/3773) [`0087dc1c0`](https://github.com/bluesky-social/atproto/commit/0087dc1c0bafad1d0a0a1a16683d250dea031bf9) Thanks [@rafaelbsky](https://github.com/rafaelbsky)! - Add verification notifications
+
+- Updated dependencies [[`0087dc1c0`](https://github.com/bluesky-social/atproto/commit/0087dc1c0bafad1d0a0a1a16683d250dea031bf9)]:
+  - @atproto/bsky@0.0.144
+  - @atproto/api@0.15.3
+  - @atproto/pds@0.4.131
+  - @atproto/ozone@0.1.106
+
+## 0.3.123
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/pds@0.4.130
+  - @atproto/bsky@0.0.143
+  - @atproto/ozone@0.1.105
+
+## 0.3.122
+
+### Patch Changes
+
+- [#3764](https://github.com/bluesky-social/atproto/pull/3764) [`30f9b6690`](https://github.com/bluesky-social/atproto/commit/30f9b6690e0e2c5810772e94e631322b9d89c65a) Thanks [@matthieusieben](https://github.com/matthieusieben)! - Use white as primary contrast color
+
+- Updated dependencies [[`30f9b6690`](https://github.com/bluesky-social/atproto/commit/30f9b6690e0e2c5810772e94e631322b9d89c65a), [`6db9faa4a`](https://github.com/bluesky-social/atproto/commit/6db9faa4a121b573fa73e29e41dd1c183543f7f8)]:
+  - @atproto/pds@0.4.129
+  - @atproto/bsky@0.0.143
+  - @atproto/ozone@0.1.105
+
+## 0.3.121
+
+### Patch Changes
+
+- Updated dependencies [[`553c988f1`](https://github.com/bluesky-social/atproto/commit/553c988f1d226b3d2fbe94c117b088f5c82db794)]:
+  - @atproto/api@0.15.2
+  - @atproto/bsky@0.0.143
+  - @atproto/ozone@0.1.105
+  - @atproto/pds@0.4.128
+
+## 0.3.120
+
+### Patch Changes
+
+- Updated dependencies [[`688268b6a`](https://github.com/bluesky-social/atproto/commit/688268b6a5ee30f0922ee152ffbd26583d164ae4), [`8d99915ce`](https://github.com/bluesky-social/atproto/commit/8d99915ce02c73b9b37bf121ccd2703fa14a906a)]:
+  - @atproto/api@0.15.1
+  - @atproto/pds@0.4.127
+  - @atproto/bsky@0.0.142
+  - @atproto/ozone@0.1.104
+
 ## 0.3.119
 
 ### Patch Changes