+++ packages/internal/handle-resolver-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto-labs/handle-resolver-node
 
+## 0.1.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto-labs/handle-resolver@0.1.8
+
 ## 0.1.14
 
 ### Patch Changes