+++ packages/oauth/oauth-client-node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atproto/oauth-client-node
 
+## 0.2.16
+
+### Patch Changes
+
+- Updated dependencies [[`0d77d1b55`](https://github.com/bluesky-social/atproto/commit/0d77d1b550a58117aee8f7f1e2be24d255ade9e4), [`30f9b6690`](https://github.com/bluesky-social/atproto/commit/30f9b6690e0e2c5810772e94e631322b9d89c65a), [`30f9b6690`](https://github.com/bluesky-social/atproto/commit/30f9b6690e0e2c5810772e94e631322b9d89c65a), [`0d77d1b55`](https://github.com/bluesky-social/atproto/commit/0d77d1b550a58117aee8f7f1e2be24d255ade9e4)]:
+  - @atproto-labs/simple-store@0.2.0
+  - @atproto/oauth-types@0.2.6
+  - @atproto-labs/did-resolver@0.1.12
+  - @atproto/oauth-client@0.3.15
+  - @atproto-labs/handle-resolver-node@0.1.15
+
 ## 0.2.15
 
 ### Patch Changes